create nav component to hold nav elements
