+++ src/CareerCoaches/App.js
@@ -1,21 +1,19 @@
 import React from 'react';
-import { useSelector, useDispatch } from 'react-redux';
+import { useSelector } from 'react-redux';
 import {
   BrowserRouter as Router,
   Switch,
   Route,
-  NavLink,
   Redirect,
 } from 'react-router-dom';
 import './App.css';
-import { logout } from '../auth/authSlice';
+import Nav from './Nav';
 import LoginPage from '../auth/LoginPage';
 import SignupPage from '../auth/SignupPage';
 import Coaches from './Coaches';
 import HomePage from './HomePage';
 
 function App() {
-  const dispatch = useDispatch();
   const loggedIn = useSelector(state => state.auth.loggedIn);
 
   if (loggedIn) {
@@ -23,25 +21,7 @@ function App() {
       <Router>
         <div className="container-fluid">
           <div className="row">
-            <div className="col-2 p-0 Nav fixed-top mr-0 px-0 h-100 text-center border-right">
-              <div className="appname">
-                <NavLink to="/">
-                  BACC
-                </NavLink>
-              </div>
-              <div className="NavLinks pt-3">
-                <NavLink to="/">Home</NavLink>
-              </div>
-              <div className="NavLinks pt-3">
-                <NavLink to="/coaches">Coaches</NavLink>
-              </div>
-              <div className="NavLinks pt-3">
-                <NavLink to="/appointments">Appointments</NavLink>
-              </div>
-              <div className="NavLinks pt-3">
-                <button type="button" className="border-0" onClick={() => dispatch(logout(false))}>Logout</button>
-              </div>
-            </div>
+            <Nav />
             <div className="col-10 offset-2 overflow-hidden bg-dark p-0">
               <Switch>
                 <Route exact path="/" component={HomePage} />
